refactor: Introduce common and atoms folders
