+++ components/common/links/InlineLink.tsx
@@ -1,5 +1,5 @@
-import Link from "next/link";
 import { cn } from "@/lib/utils";
+import Link from "next/link";
 
 export default function InlineLink({
   children,
