+++ components/common/links/NavButton.tsx
@@ -1,6 +1,6 @@
-import Link from "next/link";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
+import Link from "next/link";
 
 export default function NavButton({
   children,
