+++ components/posts/PostBody.tsx
@@ -1,6 +1,6 @@
+import { cn } from "@/lib/utils";
 import { IBM_Plex_Sans } from "next/font/google";
 import markdownStyles from "./markdown-styles.module.css";
-import { cn } from "@/lib/utils";
 
 const fontPosts = IBM_Plex_Sans({
   subsets: ["latin"],
